# Conflicts:
#	portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/controller/DashboardController.java
#	portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/controller/SprintController.java
#	portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/service/SprintService.java
#	portfolio/src/main/resources/templates/dashboard.html
#	portfolio/src/main/resources/templates/project.html
#	portfolio/src/main/resources/templates/projectForm.html
#	portfolio/src/main/resources/templates/sprintForm.html

+++ portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/controller/SprintController.java
@@ -4,7 +4,13 @@ import nz.ac.canterbury.seng302.portfolio.model.Project;
 import nz.ac.canterbury.seng302.portfolio.model.Sprint;
 import nz.ac.canterbury.seng302.portfolio.service.ProjectService;
 import nz.ac.canterbury.seng302.portfolio.service.SprintService;
+import nz.ac.canterbury.seng302.portfolio.service.UserAccountClientService;
+import nz.ac.canterbury.seng302.shared.identityprovider.AuthState;
+import nz.ac.canterbury.seng302.shared.identityprovider.ClaimDTO;
+
+import nz.ac.canterbury.seng302.shared.identityprovider.UserRole;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,20 +18,37 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import java.util.Arrays;
 import java.util.List;
 
 @Controller
 public class SprintController {
     @Autowired private SprintService sprintService;
     @Autowired private ProjectService projectService;
+    @Autowired private UserAccountClientService userAccountClientService;
 
+    /**
+     * Add project details, sprints, and current user roles (to determine access to add, edit, delete sprints)
+     * to the individual project pages.
+     * @param projectId - ID of the project selected to view.
+     * @param principal - Current User.
+     * @param model
+     * @return - name of the html page to display
+     * @throws Exception
+     */
     @GetMapping("/project/{projectId}")
-    public String showSprintList(@PathVariable("projectId") int projectId, Model model, RedirectAttributes ra) {
+    public String showSprintList(
+            @PathVariable("projectId") int projectId,
+            @AuthenticationPrincipal AuthState principal,
+            Model model,
+            RedirectAttributes ra) {
         try {
             List<Sprint> listSprints = sprintService.getSprintByProject(projectId);
             Project project = projectService.getProjectById(projectId);
             model.addAttribute("listSprints", listSprints);
             model.addAttribute("project", project);
+            model.addAttribute("roles",
+                    userAccountClientService.getUserRole(principal));
             return "project";
         } catch (Exception e) {
             ra.addFlashAttribute("messageDanger", e.getMessage());
@@ -39,7 +62,15 @@ public class SprintController {
      * @return
      */
     @GetMapping("/project/{projectId}/newSprint")
-    public String newSprint(Model model, @PathVariable ("projectId") int projectId, RedirectAttributes ra) {
+    public String newSprint(
+            Model model,
+            @PathVariable ("projectId") int projectId,
+            @AuthenticationPrincipal AuthState principal,
+            RedirectAttributes ra){
+        List<String> userRoles = Arrays.asList(principal.getClaimsList().stream().filter(claim -> claim.getType().equals("role")).findFirst().map(ClaimDTO::getValue).orElse("NOT FOUND").split(","));
+        if (!(userRoles.contains(UserRole.TEACHER.name()) || userRoles.contains(UserRole.COURSE_ADMINISTRATOR.name()))) {
+            return "redirect:/dashboard";
+        }
         try {
             Project currentProject = projectService.getProjectById(projectId);
             Sprint newSprint = sprintService.getNewSprint(currentProject);
@@ -60,40 +91,58 @@ public class SprintController {
      * @return
      */
     @PostMapping("/project/{projectId}/saveSprint")
-    public String saveSprint(@PathVariable int projectId, Sprint sprint, RedirectAttributes ra) {
+    public String saveSprint(
+        @PathVariable int projectId,
+        Sprint sprint,
+        @AuthenticationPrincipal AuthState principal,
+        RedirectAttributes ra) {
+        List<String> userRoles = Arrays.asList(principal.getClaimsList().stream().filter(claim -> claim.getType().equals("role")).findFirst().map(ClaimDTO::getValue).orElse("NOT FOUND").split(","));
+        if (!(userRoles.contains(UserRole.TEACHER.name()) || userRoles.contains(UserRole.COURSE_ADMINISTRATOR.name()))) {
+            return "redirect:/dashboard";
+            }
         try {
-            sprint.setProject(projectService.getProjectById(projectId));
-            String message = sprintService.saveSprint(sprint);
-            ra.addFlashAttribute("messageSuccess", message);
-        } catch (Exception e) {
-            ra.addFlashAttribute("messageDanger", e.getMessage());
-        }
-
-        return "redirect:/project/{projectId}";
-    }
-
-
-    /**
-     * Directs to page for editing a sprint
-     * @param sprintId ID for sprint being edited
-     * @param model
-     * @return
-     */
-    /*make sure to update project.html for path*/
-    @GetMapping("/project/{projectId}/editSprint/{sprintId}")
-    public String sprintEditForm(@PathVariable("sprintId") int sprintId, @PathVariable("projectId") int projectId, Model model, RedirectAttributes ra) {
-        try {
-            Project currentProject = projectService.getProjectById(projectId);
-            Sprint sprint = sprintService.getSprint(sprintId);
-            model.addAttribute("sprint", sprint);
-            model.addAttribute("project", currentProject);
-            model.addAttribute("pageTitle", "Edit Sprint: " + sprint.getSprintName());
-            return "sprintForm";
-        } catch (Exception e) {
-            ra.addFlashAttribute("messageDanger", e.getMessage());
+                sprint.setProject(projectService.getProjectById(projectId));
+                String message = sprintService.saveSprint(sprint);
+                ra.addFlashAttribute("messageSuccess", message);
+            } catch (Exception e) {
+                ra.addFlashAttribute("messageDanger", e.getMessage());
+            }
             return "redirect:/project/{projectId}";
         }
-    }
+
+
+
+        /**
+         * Directs to page for editing a sprint
+         * @param sprintId ID for sprint being edited
+         * @param model
+         * @return
+         */
+    /*make sure to update project.html for path*/
+    @GetMapping("/project/{projectId}/editSprint/{sprintId}")
+    public String sprintEditForm(
+            @PathVariable("sprintId") int sprintId,
+            @PathVariable("projectId") int projectId,
+            Model model,
+            @AuthenticationPrincipal AuthState principal,
+            RedirectAttributes ra){
+            List<String> userRoles = Arrays.asList(principal.getClaimsList().stream().filter(claim -> claim.getType().equals("role")).findFirst().map(ClaimDTO::getValue).orElse("NOT FOUND").split(","));
+            if (!(userRoles.contains(UserRole.TEACHER.name()) || userRoles.contains(UserRole.COURSE_ADMINISTRATOR.name()))) {
+                return "redirect:/dashboard";
+            }
+            try {
+                Project currentProject = projectService.getProjectById(projectId);
+                Sprint sprint = sprintService.getSprint(sprintId);
+                model.addAttribute("sprint", sprint);
+                model.addAttribute("project", currentProject);
+                model.addAttribute("pageTitle", "Edit Sprint: " + sprint.getSprintName());
+                return "sprintForm";
+            } catch (Exception e) {
+                ra.addFlashAttribute("messageDanger", e.getMessage());
+                return "redirect:/project/{projectId}";
+            }
+        }
+
 
     /**
      * Deletes a sprint and redirects back to project page
@@ -101,9 +150,17 @@ public class SprintController {
      * @param model
      * @return
      */
-    /* Deleting sprints */
     @GetMapping("/project/{projectId}/deleteSprint/{sprintId}")
-    public String deleteSprint(@PathVariable("sprintId") int sprintId, Model model, @PathVariable int projectId, RedirectAttributes ra){
+    public String deleteSprint(
+        @PathVariable("sprintId") int sprintId,
+        Model model,
+        @PathVariable int projectId,
+        @AuthenticationPrincipal AuthState principal,
+        RedirectAttributes ra){
+        List<String> userRoles = Arrays.asList(principal.getClaimsList().stream().filter(claim -> claim.getType().equals("role")).findFirst().map(ClaimDTO::getValue).orElse("NOT FOUND").split(","));
+        if (!(userRoles.contains(UserRole.TEACHER.name()) || userRoles.contains(UserRole.COURSE_ADMINISTRATOR.name()))) {
+            return "redirect:/dashboard";
+        }
         try {
             String message = sprintService.deleteSprint(sprintId);
             ra.addFlashAttribute("messageSuccess", message);
